fix: Maintain user identity through refreshes
We do this by maintaining a user id on the client side that is stored
in session storage so that it remains through browser refreshes.

The server uses that id to track users rather than the team nam as it
was doing previously.  In addition to the refresh problem, the team name
solution had the obvious additional problem that multiple people could
enter the same team name and cause issues.